design: 사이드바 관련 코드 삭제, 페이지 네브 버튼 헤더로 이동
